# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления курсами валют и кошельками.",
        "title": "Currency Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/rates": {
            "post": {
                "description": "Принимает значение курса в теле запроса и сохраняет его.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rates"
                ],
                "summary": "Добавить новый курс валюты",
                "parameters": [
                    {
                        "description": "Данные для создания курса (нужно только поле 'value')",
                        "name": "rate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.Rate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Курс успешно добавлен",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса или значение курса",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rates/average": {
            "get": {
                "description": "Возвращает среднее значение для последних N курсов валют.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rates"
                ],
                "summary": "Получить средний курс",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Количество последних курсов для расчета (по умолчанию 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Средний курс и количество записей",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.AverageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное значение параметра 'limit'",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets": {
            "get": {
                "description": "Возвращает массив всех зарегистрированных кошельков с их балансами.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Получить список всех кошельков",
                "responses": {
                    "200": {
                        "description": "Список кошельков",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ListWalletsResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/balance": {
            "post": {
                "description": "Создает новый кошелек с указанным балансом (если сумма положительная) или обновляет баланс существующего кошелька. Положительная сумма - пополнение, отрицательная - списание. Списание с несуществующего кошелька или до отрицательного баланса невозможно.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Создать кошелек или обновить баланс",
                "parameters": [
                    {
                        "description": "Данные для обновления баланса",
                        "name": "balance_update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.UpdateBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Баланс успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.UpdateBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса, номера кошелька или суммы",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конфликт бизнес-логики (например, недостаточно средств)",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.UpdateBalanceResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/convert": {
            "post": {
                "description": "Получает самый свежий курс, конвертирует указанную сумму и списывает ее с баланса указанного кошелька. Возвращает остаток на счете и результат конвертации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Конвертировать и списать сумму с кошелька",
                "parameters": [
                    {
                        "description": "Данные для конвертации и списания",
                        "name": "conversion_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ConvertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Конвертация и списание прошли успешно",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ConvertResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса, номера кошелька или суммы",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Указанный кошелек не найден",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конфликт: недостаточно средств на кошельке",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ConvertResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Не удалось получить актуальный курс валют",
                        "schema": {
                            "$ref": "#/definitions/currency-service_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "currency-service_internal_models.AverageResponse": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "currency-service_internal_models.ConvertRequest": {
            "type": "object",
            "properties": {
                "amount_to_convert": {
                    "type": "number"
                },
                "first_name": {
                    "description": "Пока не используется в логике, но есть в запросе",
                    "type": "string"
                },
                "last_name": {
                    "description": "Пока не используется в логике, но есть в запросе",
                    "type": "string"
                },
                "source_wallet_number": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Пока не используется в логике, но есть в запросе",
                    "type": "string"
                }
            }
        },
        "currency-service_internal_models.ConvertResponse": {
            "type": "object",
            "properties": {
                "converted_amount": {
                    "description": "Поле будет заполнено при успехе",
                    "type": "number"
                },
                "message": {
                    "description": "Сообщение об успехе или ошибке",
                    "type": "string"
                },
                "rate_used": {
                    "description": "Поле будет заполнено при успехе",
                    "type": "number"
                },
                "remaining_balance": {
                    "description": "Поле будет заполнено при успехе",
                    "type": "number"
                },
                "source_wallet_number": {
                    "type": "string"
                }
            }
        },
        "currency-service_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Содержит текст ошибки для клиента",
                    "type": "string",
                    "example": "Сообщение об ошибке"
                }
            }
        },
        "currency-service_internal_models.ListWalletsResponse": {
            "type": "object",
            "properties": {
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/currency-service_internal_models.Wallet"
                    }
                }
            }
        },
        "currency-service_internal_models.Rate": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "currency-service_internal_models.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение об успехе",
                    "type": "string",
                    "example": "Операция выполнена успешно"
                }
            }
        },
        "currency-service_internal_models.UpdateBalanceRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Может быть положительным (пополнение) или отрицательным (списание)",
                    "type": "number"
                },
                "wallet_number": {
                    "type": "string"
                }
            }
        },
        "currency-service_internal_models.UpdateBalanceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение об успехе или ошибке (например, недостаточно средств)",
                    "type": "string"
                },
                "new_balance": {
                    "type": "number"
                },
                "wallet_number": {
                    "type": "string"
                }
            }
        },
        "currency-service_internal_models.Wallet": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "Баланс кошелька",
                    "type": "number"
                },
                "number": {
                    "description": "Номер кошелька (7 знаков)",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI Spec",
        "url": "http://localhost:8080/swagger/doc.json"
    }
}